Add some checks for providers returning version and __internal (#20313)
Raised on slack due to
https://github.com/pulumi/pulumi-go-provider/pull/407

Seems go-provider was doing a couple of things that caused downstream
issues with the engine.

Firstly it was stripping the "version" field from the provider inputs in
`CheckConfig`, this resulted in the engine not storing the version in
the state file. This PR fixes this to just reset version if the provider
omits it, we also add a check that errors if a provider tries to change
the version.

Secondly the go-provider framework tries to use the property key
"__internal" but this can conflict with the engine trying to use it. We
have always just overwritten the providers data with the engines data
when this happens, but now we log a warning to tell the user it
happened.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>