Flesh out tests for dependency graphs (#19397)
The `pkg/resource/graph` module's `DependencyGraph` type is at the heart
of many Pulumi operations, allowing Pulumi to work out how changes to
one resource affect another (e.g. if a delete or replace should cascade,
or whether an update to one resource necessitates a change to another).
Unfortunately, there are a few limitations to the current implementation
of `DependencyGraph`, as well as a couple of bugs, which have been
surfaced by issues in the recent `--run-program` work. Moreover, while
historically `DependencyGraph` was designed with certain complexity
requirements in mind (avoiding quadratic-time operations where
possible), it appears that some of these properties have been lost as
the codebase has evolved. This is less of a problem in day-to-day usage,
but could potentially be affecting larger stacks/snapshots.

Before attempting to address these issues, this change makes some
changes to the test suite for `DependencyGraph` to make those changes
easier to slot in/review and reason about. It also tweaks the structure
of the test suite to make it more amenable to debugging issues with
dependency graphs in case problems or the need to extend the library
arise in future. Specifically:

* "Table"-style tests are replaced with explicit `t.Run` calls. This
makes for a slightly more verbose test suite, but allows point-and-click
"Debug this test" in VSCode (for example) to work seamlessly, making it
quick to investigate issues with correctness or performance.
* Test cases whose answers are wrong and admitting buggy behaviour are
commented as such, so that a future change which fixes them is easy to
identify/review. Existing comments about expected answers have also been
extended, and more documentation has been added about certain groups of
test cases.
* A small hacky `TestPerf` has been added to aid in measuring the
performance of a given `DependencyGraph` implementation. This does not
run by default but can be run manually. This will be referenced in a
future change to illustrate some of the complexity issues that have
crept in and to demonstrate that bug fixes/enhancements have not
introduced performance regressions (as they may have inadvertently in
the past). Note: this is consciously not a `Bench*` function that admits
a `testing.B` for reasons commented in the code.