engine: add a PluginManager interface (#20329)
These changes add a PluginManager interface to `pkg/engine` that is used
to abstract interactions with installed plugins. This allows the
lifecycle test framework to supply an implementation of the plugin
manager that does no I/O and always treats all plugins as installed.
This
could also be used to improve test coverage for
`EnsurePluginsAreInstalled`.