improve snapshotEqual algorithm (#20323)
This algorithm is currently O(n^2), where n is the number of
resources/pending operations. For the lifecycle tests where we currently
use it this doesn't really matter, as we always have very few resources.

However we want to start using this algorithm to compare the final
snapshots, between the journaling snapshot manager, and the regular
snapshot manager. In this case snapshots can get quite large, and
quadratic behaviour would not be ideal. Not sure if we actually see
large enough snapshots in practice for this to matter, but the changes
is relatively simple, so it's better to be safe than sorry.

The change here is that we first start building a map of urn to
resource/pending operation. That means that for each resource/pending
operation in the journal, we only need to run through a very limited
amount of resources in the regular snapshot managers to compare them to.

In practice, in an example with 5000 resources, this helps save 3
minutes.

Current algorithm:

> `pulumi up --yes  1338.37s user 29.04s system 156% cpu 14:34.10 total`

Proposed algorithm:

> `pulumi up --yes  1064.58s user 26.34s system 162% cpu 11:13.12 total`