[convert] Warn instead of erroring out when traversing a NoneType in PCL (#20342)
Fixes #20322

There are many scenarios (especially in TF) where a program tries to
index an empty object. The converted PCL from that errors out because
traversing a `NoneType` isn't _supported_. The result of the type
traversal itself returns a `NoneType` still. I thought of two solutions:
- return `DynamicType` when traversing `NoneType` without returning a
diagnostic
 - make the returned diagnostic a warning instead of an error (this PR) 

I went for the latter because it doesn't change the semantics of the
program although it might a good idea later to make traversals more
lenient and return `DynamicType`