[PCL] Simplify range value types when it is derived from object of objects and warn on unknown properties (#20345)
Fixes the type-checking issue of
https://github.com/pulumi/pulumi-converter-terraform/issues/373 making
CLI more resilient when encountering programs of that try to traverse an
object of objects and then indexing a non-existent key 😓

Part of the problem is that when traversing an object of objects shaped
like this
```
accounts = {
  "us-east-1" = {
    awsRegion = "us-east-1"
    something = 10
  }
  "us-west-2" = {
    awsRegion = "us-west-2"
    something = 20
  }
  "eu-west-1" = {
    awsRegion = "eu-west-1"
    something = 30
  }
}
```
When such object is used in a `range` expression of a resource, then we
try to extract the `rangeKey` and `rangeValue` types of the collection
being iterated. In the case of the above the `rangeValue` is computed to
be a union of all the objects types in the map `union(object1, object2,
...., objectN)`

This PR simplifies this process by instead making `rangeValue` a single
object where all the properties are unified instead.

Also treat unknown property errors into warnings during resource binding
if we are binding in non-strict mode