Disambiguate duplicate resource names by their types when importing (#20348)
I think this fixes #20318 because code generation in PCL does not handle
declaring multiple nodes (in this case resources) of the same name.

This PR tries to disambiguate the names of the resources to be generated
in the `NamesTable` using their resource type.

```
{
	Name: "thing",
	ID:   "thing",
	Type: "pkg:module/Baz:First",
},
{
	Name: "thing",
	ID:   "thing",
	Type: "pkg:module/Quux:Second",
},
```
generates this `NamesTable`
```go
importer.NameTable{
	"urn:pulumi:stack::proj::pkg:module/Baz:First::thing":   "thing",
	"urn:pulumi:stack::proj::pkg:module/Quux:Second::thing": "thing_Second",
}
```
which is then used in generated PCL.

It is possible that there are multiple resources of the same name & type
from a different module or package that this code doesn't handle but
that seems unlikely