Fix dependency graph hangs in refresh  (#20339)
Fixes https://github.com/pulumi/pulumi/issues/20274

This fixes an infinite hang we could get in the dependency system due to
passing the `refreshStates` map values in without sorting them. Doing
that could result in `NewDependencyGraph` looping forever as it tried to
lookup the parent for the first resource, found no match in the
`urnIndex` map, but then because Go just returns the zero value for
missing values, it would then loop and try and do the first resource (at
index 0) again, and again.

Turns out we don't even need to pass the values from the `refreshStates`
map into the dependency calculation, we just need to alias them, the
same we do for the top-level dependency graph stored on the deployment.

Tested this with the repro in #20274 and the refresh step finishes
without issue.