try to unflake TestResourceRefsGetResourceNode test
This test was re-enabled in
https://github.com/pulumi/pulumi/pull/16903, but is still flaky.  CI
can be very slow sometimes, so let's increase the time we're waiting
for outputs, in hopes that that unflakes the test.

While there also fix up tabs vs. spaces, as my original PR introduced
a bunch of tabs where the indentation should be with spaces.