[testing] Fix TestGoModEdits for alt worktrees (#20366)
These changes fix TestGoModEdits to account for the fact that the
worktree for the `github.com/pulumi/pulumi` repository may not be
located at a path that terminates in `pulumi`. In that case,
`getEditStr` resolves the replacement path to different locations
depending on whether or not the module being replaced uses a relative
path.