Set ThreadPoolExecutor max_workers to parallelism * 4 (#20362)
The GRPC monitor client uses blocking operations, so we run them in a
ThreadPool. This was configured to use the value of the `--parallel`
flag, which controls the engine parallelism. However other operation
than just resource registrations can occupy slots in the threadpool, and
reduce the number of calls the SDK could make.

This could also lead to deadlocks when `--parallel` was set to a low
value and a transform would call for example an invoke. In this case a
resource registration would block until the transform finishes, but the
transform would block waiting to get a thread to run the invoke. Neither
of the threads get freed up, and we are blocked.

Set the max_workers for the pool to parallelism * 4 to give us more head
room here.