Fix TestRefreshInitFailure to not just delete all the resources (#20360)
This test was pretty useless because it looped over and asserted
properties of the resources, but didn't assert that the resources _were_
in the list to assert against.

This fixes the test to return ID so refresh doesn't just delete the
resources. I'm pretty sure there's more issues here if we turn on
`WithGrpc` but this is a start.