introduce a validating persister that validates the snapshot (#20363)
Introduce a validating persister, that validates the snapshot each time
it is persisted, iow, when `Save` is called. It works using a callback
that can collect the errors and do with them as the caller desires.

The idea is to use this for the journaling code to verify the snapshot
after each step. Introduce it here in a separate PR to keep
https://github.com/pulumi/pulumi/pull/19862 from growing too much.