Add `previewDestroy` to the NodeJS SDK (#20371)
Continuing in the steps of #19900, #19948, and so on, this PR adds the
`previewDestroy` command and deprecates the `--preview-only` flag in the
`destroy` command. The reason is that `destroy` (like `refresh`) uses
the event log to figure out what was affected by a change. However,
preview commands don't generate event logs, so you instead get the
results of whatever the _previous_ change was. This PR special-cases
this preview behaviour to avoid the bug.