Default lifecycle Read methods to be StatusOK (#20374)
We're filling in all the inputs and outputs as if nothings changed,
which will be fine for most of the tests. As such we can just. return OK
here instead of Unknown, I don't think any test is affected by this yet
but I could see subtle changes causing Unknown to start causing issues.