Fixup old inputs/outputs for DiffConfig to contain version
In https://github.com/pulumi/pulumi/pull/20313 we changed the engine to ensure
"version" was always added to provider inputs that are saved to state. This had
the effect of also meaning those inputs were passed to `DiffConfig` with
"version" set, while previously if the provider had been stripping it in
`CheckConfig` it wouldn't see it in `DiffConfig`.

If a provider didn't implement DiffConfig this would then hit the engines
default logic of doing a simple compare of the new state (with version) against
the old state (without version) and report a provider diff. A provider
implmenting it's own version of this logic in `DiffConfig` would see the same
thing because the old inputs and outputs it was being passed we're likewise
missing "version".

This fixes up the registry to restore "version" to the old inputs/outputs if
it's missing. We fill in from the version found on the new inputs (which again
due to #20313 will be there). This isn't technically correct, we don't know
what the old version of the state is, but assuming its the same version and
passing something looks like it will give us better real-world behaviour than
continuing to not pass "version" and hitting issues like
https://github.com/pulumi/pulumi-docker-build/issues/581.