fix flaky TestRangeTraversalFromObjectOfObjectsDoesNotError
This test is sometimes flaky because the array of unknown properties
is not sorted, and can appear in different orders.

The test fails with:

```
        	Error:      	"unknown property 'aws_region' among [something awsRegion]" does not contain "unknown property 'aws_region' among [awsRegion something]"
```

We go for the simple fix here, just sorting that array, as it should
always be cheap enough to do so, and is the best way to fix the test.