Limit @types packages used during local SDK generation (#20387)
Generated local SDKs depend directly on typescript because we require
the ts compiler to build the package in a postinstall script. The
dependency ensures that the compiler is present and at a version that is
guaranteed to work for the local SDK.

By default typescript will include any `@types/…` packages. When the
project that uses the local SDK includes a `@types/…` package that uses
features requiring a more recent typescript compiler, the postinstall
script will fail. An example is `@types/jest@30`, which requires
typescript 5.2, but local SDKs use 4.3.

To fix this we can pass the `--types` flag to `tsc` to select which
`@types/…` packages are included in the compilation and limit it to only
those the local SDK depends on, not those of the project.

Fixes https://github.com/pulumi/pulumi/issues/20386