Implement `pulumi state taint/untaint` commands (#20380)
This change implements https://github.com/pulumi/pulumi/issues/11657,
which is a new set of commands to "taint" and "untaint" resources.
Similar to Terraform's like-named command, this just sets a bit in the
statefile that is used to indicate that the resource should be replaced
upon the next `pulumi up`. This leverages the machinery we already have
in place with the `pulumi up --replace` command, and works similar to
its siblings `pulumi state protect/unprotect` in the UX (i.e., how URNs
are passed and/or chosen interactively). I didn't think `pulumi taint
--all` made sense (but open to it), although `untaint --all` does.

Fixes #11657

---------

Co-authored-by: Fraser Waters <fraser@pulumi.com>
Co-authored-by: Justin Van Patten <jvp@justinvp.com>
Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>