fix flaky TestRangeTraversalFromObjectOfObjectsDoesNotError (#20407)
This test is sometimes flaky because the array of unknown properties is
not sorted, and can appear in different orders.

The test fails with:

```
        	Error:      	"unknown property 'aws_region' among [something awsRegion]" does not contain "unknown property 'aws_region' among [awsRegion something]"
```

We go for the simple fix here, just sorting that array, as it should
always be cheap enough to do so, and is the best way to fix the test.

noticed in
https://github.com/pulumi/pulumi/actions/runs/17270052509/job/49012394297?pr=20403.

Fixes https://github.com/pulumi/pulumi/issues/20406