[chore] Replace NewState with NewState.Make
These changes replace `resource.NewState` with an args type. This type
exports a single method, `Make`, that retuns a fresh `State` value.

This pattern was chosen over making all fields of `State` required
because it is common (and harmless) for tests to directly create
sparsely-specified `State` values. Another option would be to make all
fields of `State` required and add a second type that is another name
for `State` (e.g. `SparseState`, `TestState`, etc). That approach would
allow us to remove `NewState` entirely at the cost of updating all of
our test code to use the renamed type.