update mise guide to show enabling experimental mode (#20426)
The mise configuration uses Go package installation, this is still
experimental and requires a flag to be enabled:
https://mise.jdx.dev/dev-tools/backends/go.html

I've updated the documentation to show how to enable it either with a
setting or an environment variable. I'll keep an eye on if this changes
and remove it later.

Question: 
Happy with me not linking the reference here?

---------

Co-authored-by: Pat Gavlin <pgavlin@gmail.com>