Handle missing cases in `ConvertJSONEvent` (#20429)
`ConvertJSONEvent` is used when running a remote deployment via
automation API (via the CLI), to convert the engine events being
produced by the remote deployment into engine events that can be
displayed locally.

I was looking at the recent change that added `ErrorEvent` and noticed
it wasn't being handled here. It's not a problem yet since the CLI is
not producing this event yet, but could turn into a problem when it
does. This change handles `ErrorEvent` and adds a test that ensures
we're exhaustively handling all possible cases on `apitype.EngineEvent`,
since we can't lean on a linter in this case.

The test found that we're also not handling `StartDebuggingEvent`. It's
arguable if we even need to handle this internal event here, but it is a
field on `apitype.EngineEvent` and we are handling the ephemeral
`ProgressEvent` event, (and it's simple enough to handle), so I've
chosen to do so.

As a possible follow-up, we may want to reconsider whether we should
have internal/ephemeral events on `apitype.EngineEvent`.