Add policy summary events
Add `PolicyAnalyzeSummaryEvent`, `PolicyRemediateSummaryEvent`, and `PolicyAnalyzeStackSummaryEvent` events that summarize which policies passed, failed, and were not applicable for each `Analyze`, `Remediate`, or `AnalyzeStack` call.

This is the first step of introducing these new events. Next will be consuming the new events from the service, followed by actually emitting the events from the CLI.