improve usage string for pulumi config set (#20437)
In `pulumi config set`, `--` can be used to disambiguate values that
look like flags, iow, they have the format `--<value>`. Make this clear
in the usage string. Note that this means flags should be set before the
`--`, so move the `[flags]` part earlier in the usage string as well.

Fixes https://github.com/pulumi/pulumi/issues/20434