Fix not using already-installed package when PULUMI_EXPERIMENTAL=1
If a package is already installed, we should always default to using it instead of resolving to some other package. Otherwise, in this sequence:

```
e.RunCommand("pulumi", "plugin", "install", "resource", "terraform-provider", "0.3.1")
e.RunCommand("pulumi", "package", "add", "terraform-provider", "hashicorp/random", "3.6.3")
```

The second command will run the package resolver and try to install the latest-available version of `terraform-provider`, which may not be `0.3.1` (in the case of our tests, it was `0.13.0`).

Fixes https://github.com/pulumi/pulumi/issues/20441

Should hopefully unblock #20361 after this merges and we rebase.