Use the passed through writer in ShowProgressOutput instead of loading a new terminfo
Instead of loading terminfo from `$TERM`, check if the writer is a terminal and use that to determine if we should print ansi escapes.

This ensures that we use the MockTerminal during tests, which prints placeholders for the escapes, such as `<{%underline%}>` or `<%clear-to-cursor%>`. This makes the golden files consistent across all operationg systems, including Windows. The test also does not need to set `$TERM` anymore.