[proto] Analyzers: Return "not applicable" policies
Update `AnalyzeResponse` and `RemediateResponse` with support for returning a list of "not applicable" policies, and update `GetAnalyzerInfo` with support for returning the type of policy (resource or stack) for each policy. This way, for a given call to `Analyze`, `Remediate`, or `AnalyzeStack`, we can determine which policies failed, passed, and were not applicable.