Replace terminfo package (#20446)
The `github.com/ijc/Gotty` package doesn't know how to load the terminfo
file on macOS (or FreeBSD), so we were always using the fallback paths
on these systems (and on Windows, because terminfo is not a thing
there).

The tests run with `TERM=vt102`, and the escape codes for some
operations are different from those in the fallback paths, so the golden
files are inconsistent. Now that we can correctly load the terminfo
file, we get consistent behaviour in the tests when running on Linux and
macOS.

Ref https://github.com/pulumi/pulumi/issues/19675

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>