propagate context through cancellationScopeSource (#20421)
WE have a `cancellationScopeSource` that we use for terminating the
program using `SIGINT`. Currently we generate a new context for this
purpose. However I think it makes more sense to pass in a context, so
the cancelation scope is also canceled if the parent context is
canceled.

Note that if we call `cancel` on the `cancelContext` inside the
cancellation source, the parent context will not be canceled, so there's
no change in behaviour there.

I looked into this because of the conversation here:
https://github.com/pulumi/pulumi/pull/20397#discussion_r2301734743.