Use the passed through writer in ShowProgressOutput instead of loading a new terminfo (#20449)
Instead of loading terminfo from `$TERM`, check if the writer is a
terminal and use that to determine if we should print ansi escapes.

This ensures that we use the MockTerminal during tests, which prints
placeholders for the escapes, such as `<{%underline%}>` or
`<%clear-to-cursor%>`. This makes the golden files consistent across all
operationg systems, including Windows. The test also does not need to
set `$TERM` anymore.

Since now there are no golden files that record any ansi escape codes,
this also adds a small test to make sure the basic mechanism to lookup
codes works as expected.

Ref https://github.com/pulumi/pulumi/issues/19675

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>