Fix the --show-full-output the flag
We have a `--show-full-output` flag (default `true`) that allows truncating long outputs in the diff display, e.g. `pulumi up --diff`. Pass this through to the progress renderer (the default display) so that long output lines are truncated.

The truncate option was slightly buggy and not properly truncating based on newlines. The tests also did not set the truncate option where they should.

```
Outputs:
    a     : [secret]
    b     : "123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123123..."
```

This option was added in https://github.com/pulumi/pulumi/pull/10021 for the diff display