Use built-in protobuf Struct / Go map marshaling code
We had to have our own marshaling code to convert protobuf Structs to/from Go maps because there wasn't anything built-in to do this when this code was originally written. But since then, built-in helpers were added to google.golang.org/protobuf with https://go-review.googlesource.com/c/protobuf/+/225298. This change switches over to use the built-in functionality.