[program-gen] Fix panic in RewriteApplies when encountering warning diagnostics (#20453)
Fixes #20419

Starting from Pulumi CLI v3.191.0 an example conversion in AWS upgrade
started panic'ing where as previously it didn't.

Looking into changes PCL in v.3.191.0 we merged the following PRs

- [programgen] Emit warnings instead of errors in non-strict mode when
binding resource blocks
https://github.com/pulumi/pulumi/pull/20347

- [programgen] Simplify range value types when it is derived from object
of objects and warn on unknown properties
https://github.com/pulumi/pulumi/pull/20345

- [programgen] Warn instead of erroring out when traversing a NoneType
in PCL
https://github.com/pulumi/pulumi/pull/20342

After investigating, undoing each of the PRs separately it turns out
that the first PR https://github.com/pulumi/pulumi/pull/20347 is the
culprit.

The reason it was panic'ing when converting to a specific language is
that before https://github.com/pulumi/pulumi/pull/20347, we would error
out at the bind/typecheck phase and not generate any program. Now with
https://github.com/pulumi/pulumi/pull/20347, we turned error diagnostics
(in non-strict mode) into warning diagnostics so that program generators
can handle the rest. At least that was the assumption, turns out that
when we are generating a program and we run `RewriteApplies(...)`
against expressions in the resource blocks, they cannot have any type
errors (e.g. indexing a non-existent property) otherwise it violates a
contract in the implementation of `RewriteApplies`

This PR fixes the issue by making the contract assert against _error_
diagnostics rather than no diagnostics at all, thus allowing for warning
diagnostics to pass without panic'ing. Also it teaches
`model.ObjectType` about strict mode so that typechecking and traversals
don't emit errors but warnings instead