journaling interface inside the engine (#19862)
This PR changes the internal snapshot interface to use journal entries
instead of mutating the snapshot at every step. There's currently no
functional changes intended, as we still do snapshotting the same way as
before, creating a new snapshot as each journal entry is processed in
sequence.

In future steps we will be able to send the journal entries directly to
the service, which can then reconstruct a snapshot from them, instead of
having to send the whole snapshot back and forth. These journal entries
can be sent to the service asynchronously, as the ordering doesn't
matter, because every entry encodes a particular change, instead of
encoding the whole snapshot that can't be overwritten later.

Part of https://github.com/pulumi/pulumi/issues/13502

---------

Co-authored-by: Julien <julien@caffeine.lu>