Add a public export map accessor for unit testing (#20471)
Part of #6113.

It's nice to be able to write a unit test that runs your program to
produce some exports, and then verify the contents of those exports.
Currently, there's no way to see the created exports after the fact as
the export map is private. This PR creates a public accessor for this
map for the purposes of unit testing.