[engine] Journal tweaks
- Add some helper functions to the journaler
- Don't send Begin entries for sames, refreshes, removes
- Rename OperationID to EndOperation and make it optional
- Rename SequenceID to SequenceNumber

With respect to not sending begin entries: the original snapshotter did
not write pending operations for these operations, and the begin entry is
not necessary in order to update any of the state.

Regarding the rename from OperationID to EndOperation: the only time we
need to correlate between journal entries is when removing a pending
operation from the statefile. We can use the sequence number of the
entry that added the operation to perform this correlation.