[engine/sdk] Add support for resource stack traces (#20430)
These changes add support for associating resource states with the
contents of
the Pulumi program's call stack at the time the resource was registered.
The
details of the changes to each component are below, starting with the
engine.

# Engine

These changes add the necessary machinery to the engine to capture and
persist
stack traces. This includes support for concatenating stack traces that
cross
plugin boundaries.

Capturing and persisting raw stack traces themselves is pretty simple.
Resource
monitor clients send stack traces to the monitor as part of Read,
Register, and
Call, after which the engine need only persist the traces in state.

Stack concatenation requires that the resource monitor is able to track
call
stacks across plugins. Call stacks may cross plugins as part of a call
to Call
or the registration of a remote component. In both cases, the monitor
extracts
the incoming stack from the request and associates it with a handle
(currently
a UUID, but could be anything sufficiently unique). This handle is then
passed
on to the associated Call or Construct call made by the monitor in
response to
the original request. When the monitor receives a Call/Read/Register
that
contains a parent stack handle, it extracts the handle's associated
stack trace
and prepends it to the stack trace from the request.

# Protocol

The following requests may now contain stack traces from the caller:

- ReadResourceRequest
- RegisterResourceRequest
- ResourceInvokeRequest
- ResourceCallRequest

All of the above are sent by resource monitor clients. Their stack
traces may
reference code in the user's Pulumi program or code in a multi-language
component in a provider. In addition to the caller's stack trace, all of
the
above besides ResourceInvokeRequest may also contain a handle for a
parent
stack trace. That handle is used by the engine to concatenate stacks for
calls
that cross plugin boundaries.

The following requests may now contain a handle to a stack trace:

- CallRequest
- ConstructRequest

These requests are sent to providers by the engine. The stack trace
handle
should be recorded by the provider so that subsequent requests to the
resource
monitor can contain the handle for stack concatenation.

# Statefiles

Stack traces are captured in state files using two new types: one for
the stack
trace and one for each frame it contains. This will allow us to add
additional
information to traces and frames in the future if we so choose.

Note that the "sourcePosition" property of a stack frame is optional.
This
allows us to support stacks that contain frames that were missing
position
information at the time they were captured.

# SDKs

Each of the Go, NodeJS, and Python SDKs has been updated to send stack
traces
to the engine as part of resource registration. **NOTE**: the SDKs have
not yet
been updated to send stack traces for Call, and do not yet support stack
handles for stack concatenation. These changes are somewhat more
invasive, and
will be added as a follow-up.