Skip update check on command init (#20502)
When initializing the PulumiCommand object, we need to load the CLI
version.

By default, this will also perform an update check and print a warning
if out of date. In this context, the warning will not be displayed, but
the addional network call will slow down construction of workspaces etc.

Therefore, this is a small optimization to disable the update check for
this command.

---------

Co-authored-by: Fraser Waters <fraser@pulumi.com>