pass context through to ProjectInfoContext (#20397)
We previously created a new context.TODO() inside this function, while
we should be passing through the real context from outside. I noticed
this in particular because opentelemetry uses the context for parenting
spans correctly. However it seems like the right thing to do either way
to pass the context through here.