Fix parents in python components (#20508)
The server code for python components had it's if condition flipped for
applying parents. That is if the URN was non-empty it ignored it, and if
it was empty it tired to construct a dependency resource with an empty
URN.

This fixes the condition and changed the component tests to show
parenting works.

Fixes https://github.com/pulumi/pulumi/issues/20507