make journal replayer shareable
We want to re-use the journal replayer in the service.  Since the
service only operates on apitypes, make the journal replayer also work
that way.  This means serializing journal entries for ingestion, and
deserializing the snapshot for comparison locally.  Once we use
journaling in the service as well, this isn't going to be any more
overhead, because we need to serialize the journal entries anyway, and
won't need to deserialize the deployment anymore.