make roundtripping serialized resources with secrets work
When serializing resources with secrets, we use the
`SerializePropertyValue` function, which internally uses
`apitype.SecretV1` to represent secrets.  However the corresponding
`DeserializePropertyValue` function only supports
`map[string]any`. This works currently because we usually roundtrip
these property values through the backend, and deserializing from JSON
we get the `map` value back, even when we originally serialized a
`SecretV1`.

In subsequent PRs we want to introduce serializing and deserializing
of resources within the CLI, so deserializing a `SecretV1` needs to
work. Make that so. The alternative would be to just return a map from
`SerializePropertyValue`, but that seems worse than using the proper
types where we can.

The serializing and deserializing within the CLI comes from the desire
to share the journal replay logic with the service. For this it needs
to work with serialized journal entries, but then to do comparisons on
the deployment we need to deserialize it.