copy new resources as well as old ones
Currently in the journal implementation we only made copies of the old
snapshot, but then added new resources without copying them.  This
meant that the engine was still updating pointers, which is not what
we want, as that doesn't translate to the service.

Make the journal make copies of new entries as well to avoid this.