Unflake TestRunningViaCLIWrapper (#20527)
When cancelling an operation, we send SIGINT to plugins but don’t wait
for the processes to exit. This makes the test flaky. To fix this, we
wait for the `interrupted.txt` file to be written (wtih a timeout). This
tells us that the provider recieved the signal and wasn’t stuck.