generate skipped create steps for providers during refresh
During `refresh --run-program`, we currently don't handle provider
resources specially, unlike other custom resources. For custom
resources that are new, we create "skipped creates", which don't
actually go into the snapshot.

However for providers, or more specifically default providers we can't
do that. Refreshed resources might need a new default provider, so we
have to create those providers properly.

This falls apart when we try to create a provider that has a
dependency, in particular a dependency on a resource that's new in the
program, but isn't actually getting created by the refresh. Because we
try to create regular create step, we also end up trying to write that
to the snapshot.  But since the dependency is missing, we get a
snapshot integrity error.

Providers that have dependencies are always explicit providers. Issue
skipped creates for such providers. This means everything depending on
them will also fail to be created.

Co-authored-by: Raj Perera <rajiteh@gmail.com>