expand snapshot integrity check to check that resources are well formed (#20519)
When [deserializing
resources](https://github.com/pulumi/pulumi/blob/f5953999c151ea354a448c5f641c5cc37b33d2f3/pkg/resource/stack/deployment.go#L651-L661),
we currently check some properties of resources, e.g. that it has a URN,
a type, and that component resources have no ID.

However we currently don't check the same when verifying snapshot
integrity. This means we could pass a snapshot integrity check, and send
the snapshot to the backend, but could then fail to derserialize the
snapshot. Fix this by expanding the snapshot integrity check, and fixing
up the tests that violate these checks.

I'm a little scared of the `VerifyIntegrity` changes, since it could be
that we get more snapshot integrity errors from this, although I *think*
users should already be getting deserialization errors if that was the
case. The test changes should be merged either way imo, and are going to
be necessary for a future PR I'm working on.