Fix plugin killing during cancellation
`TestRunningViaCLIWrapper` tests cancellation via Ctrl-C/SIGINT. The test was broken in two ways:

* `pulumi package add ...` runs the component provider and then shuts it down normally. This sends a SIGINT to the plugin, so it creates the `interrupted.txt` sentinel file.
* The program did not actually use the provider, so during the update operation we didn’t start the component provider.

Fixing the test by removing the `interrupted.txt` created during `pulumi package add`, and creating a component in the program revealed a bug in how we shutdown plugins.

Plugins are first sent a SIGINT. If the plugin does not terminate within 5 seconds, we send a SIGKILL. However if the operation is cancelled, the `context.Context` created for the 5 second timeout was parented to a cancelled context, so was also cancelled. This meant we never sent a SIGKILL.

Usually this is not a problem, but because this provider implements a SIGINT handler, it did not have the deafult SIGINT behaviour of exiting, and would stay running. In this case we needed the SIGKILL fallback, but that didn’t work because of the cancelled context.